Move RFQT Mocker to src
