+++ packages/asset-swapper/package.json
@@ -55,6 +55,7 @@
         "@0x/utils": "^5.4.1",
         "@0x/web3-wrapper": "^7.0.7",
         "axios": "^0.19.2",
+        "axios-mock-adapter": "^1.18.1",
         "heartbeats": "^5.0.1",
         "lodash": "^4.17.11"
     },
@@ -72,7 +73,6 @@
         "@types/lodash": "4.14.104",
         "@types/mocha": "^5.2.7",
         "@types/node": "*",
-        "axios-mock-adapter": "^1.18.1",
         "chai": "^4.0.1",
         "chai-as-promised": "^7.1.0",
         "chai-bignumber": "^3.0.0",
